requirements.txt
cloudinary
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
environ==1.0
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
packaging==20.9
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
serializers==0.2.4
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
